Move tests for deprecated functionality to `UnitDeprecated`
