+++ tests/Unit/Position/PositionTest.php
@@ -191,94 +191,4 @@ final class PositionTest extends TestCase
         self::assertSame($lineNumber, $this->subject->getLineNumber());
         self::assertSame($columnNumber, $this->subject->getColumnNumber());
     }
-
-    // Testing deprecated functionality
-
-    /**
-     * @test
-     */
-    public function getLineNoInitiallyReturnsZero(): void
-    {
-        self::assertSame(0, $this->subject->getLineNo());
-    }
-
-    /**
-     * @test
-     *
-     * @dataProvider provideLineNumber
-     */
-    public function getLineNoReturnsLineNumberSet(int $lineNumber): void
-    {
-        $this->subject->setPosition($lineNumber);
-
-        self::assertSame($lineNumber, $this->subject->getLineNo());
-    }
-
-    /**
-     * @test
-     */
-    public function getLineNoReturnsZeroAfterLineNumberCleared(): void
-    {
-        $this->subject->setPosition(99);
-
-        $this->subject->setPosition(null);
-
-        self::assertSame(0, $this->subject->getLineNo());
-    }
-
-    /**
-     * @test
-     */
-    public function getColNoInitiallyReturnsZero(): void
-    {
-        self::assertSame(0, $this->subject->getColNo());
-    }
-
-    /**
-     * @test
-     *
-     * @dataProvider provideColumnNumber
-     */
-    public function getColNoReturnsColumnNumberSet(int $columnNumber): void
-    {
-        $this->subject->setPosition(1, $columnNumber);
-
-        self::assertSame($columnNumber, $this->subject->getColNo());
-    }
-
-    /**
-     * @test
-     */
-    public function getColNoReturnsZeroAfterColumnNumberCleared(): void
-    {
-        $this->subject->setPosition(1, 99);
-
-        $this->subject->setPosition(2);
-
-        self::assertSame(0, $this->subject->getColNo());
-    }
-
-    /**
-     * @test
-     */
-    public function setPositionWithZeroClearsLineNumber(): void
-    {
-        $this->subject->setPosition(99);
-
-        $this->subject->setPosition(0);
-
-        self::assertNull($this->subject->getLineNumber());
-    }
-
-    /**
-     * @test
-     */
-    public function getLineNoAfterSetPositionWithZeroReturnsZero(): void
-    {
-        $this->subject->setPosition(99);
-
-        $this->subject->setPosition(0);
-
-        self::assertSame(0, $this->subject->getLineNo());
-    }
 }
